Delete old k3d lib
